app/requirements.txt
fastapi==0.115.8
redis==5.2.1
uvicorn==0.34.0
SQLAlchemy==2.0.37
pydantic-core==2.23.4
pydantic-settings==2.7.1
pydantic==2.9.2
redis==5.2.1
psycopg2==2.9.10
oauth2client==4.1.3
yandexcloud==0.334
gunicorn==20.1.0
alembic==1.13.1
PyMuPDF==1.25.5
python-multipart==0.0.20
langchain==0.3.17
langchain-core==0.3.58
langchain-community==0.3.16
langchain-google-genai==2.1.4
langchain-openai==0.3.16
debugpy==1.8.14
socksio==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client>=2.0.0,<3.0.0
google-auth>=1.16.0,<3.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
langgraph
rapidfuzz